suit_research/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database - PostgreSQL
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Database - MongoDB
motor==3.3.2
pymongo==4.6.0

# Cache - Redis
redis==5.0.1
aioredis==2.0.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Monitoring and Logging
structlog==23.2.0

# Web scraping (for crawlers)
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
urllib3==2.1.0

# Rate limiting
slowapi==0.1.9

# Additional utilities
click==8.1.7